requirements.txt
fastapi
uvicorn
pydantic
numpy
PyYAML
requests
openai
ollama
spacy==3.5.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0.tar.gz
streamlit
sqlite-utils
sqlite-vec
PyPDF2
langchain
langchain-community
tiktoken
langchain_experimental
json-repair
google-auth
google-cloud-aiplatform
langchain-google-vertexai
langgraph
langchain-chroma
faiss-cpu